+++ game-frontend/src/app/display-single-upgrade/display-single-upgrade.component.ts
@@ -3,8 +3,11 @@ import { RunningUpgrade } from './../shared-pojo/running-upgrade.pojo';
 import { UpgradeService } from './../service/upgrade.service';
 import { ObtainedUpgradePojo } from './../shared-pojo/obtained-upgrade.pojo';
 import { MEDIA_ROUTES } from './../config/config.pojo';
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter, ViewChild } from '@angular/core';
 import { Upgrade } from './../shared-pojo/upgrade.pojo';
+// tslint:disable-next-line:max-line-length
+import { WidgetConfirmationDialogComponent } from '../modules/widgets/components/widget-confirmation-dialog/widget-confirmation-dialog.component';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-display-single-upgrade',
@@ -22,11 +25,13 @@ export class DisplaySingleUpgradeComponent extends BaseComponent implements OnIn
   @Output()
   public onRunningUpgradeDone: EventEmitter<{}> = new EventEmitter();
 
+  @ViewChild('confirmDialog') public confirmDialog: WidgetConfirmationDialogComponent;
+
   public image: string;
-
   public runningUpgrade: RunningUpgrade;
+  public vConfirmDeleteText: string;
 
-  constructor(private _upgradeService: UpgradeService) {
+  constructor(private _upgradeService: UpgradeService, private _translateService: TranslateService, ) {
     super();
     this.requireUser();
   }
@@ -44,8 +49,26 @@ export class DisplaySingleUpgradeComponent extends BaseComponent implements OnIn
     this._upgradeService.registerLevelUp(selected);
   }
 
-  public cancelUpgrade(): void {
-    this._upgradeService.cancelUpgrade();
+
+  /**
+   *
+   * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+   * @since 0.7.2
+   * @returns {Promise<void>}
+   * @memberof DisplaySingleUpgradeComponent
+   */
+  public async clickCancelUpgrade(): Promise<void> {
+    this.vConfirmDeleteText = await this._translateService.get(
+      'UPGRADE.CANCEL_TEXT',
+      { upgradeName: this.upgrade.name }
+    ).toPromise();
+    this.confirmDialog.show();
+  }
+
+  public cancelUpgrade(result: boolean): void {
+    if (result) {
+      this._upgradeService.cancelUpgrade();
+    }
   }
 
   public otherUpgradeAlreadyRunning(): void {