+++ game-frontend/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 v0.7.2 (latest)
 ===========================
+* __Improvement:__ Display a cancel confirmation in upgrades  
+![Display a cancel confirmation in upgrades](assets/changelog/improvements/kBMYJtvv.png)  
 * __Fix:__ Close icon not displaying in modals
 * __Fix:__ Sometimes units don't appear, but, it's counting the number of soldiers
 * __Fix:__ Register and forgot password links are inverted